% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_bar_chart_labels}
\alias{get_bar_chart_labels}
\title{Bar chart function}
\usage{
get_bar_chart_labels(data, variable, title = "Plot_title")
}
\arguments{
\item{data}{A labelled dataset}

\item{variable}{The variable you want to plot}

\item{title}{The name of the plot}
}
\value{
A ggplot bar plot graph
}
\description{
This function produces a ggplot bar chart with the appropriate labels that shows the percentage for each level.
}
