% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_bar_plot_index}
\alias{get_bar_plot_index}
\title{Answer index graph function}
\usage{
get_bar_plot_index(data, grouping_var, title = "Plot_title")
}
\arguments{
\item{data}{A dataset with the distribution of answers per level of the grouping variable}

\item{variable}{The variable you want to plot}
}
\value{
A ggplot bar plot graph
}
\description{
This function allows you to produce a ggplot graph plot
}
