% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_bar_simple}
\alias{get_bar_simple}
\title{Simple bar chart function}
\usage{
get_bar_simple(data, variable)
}
\arguments{
\item{data}{A dataset}

\item{variable}{The variable you want to plot}
}
\value{
A ggplot bar plot graph
}
\description{
This function allows you to produce a ggplot bar char plot without axis levels. Useful for plotting several variables in a single grid.
}
