% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_index}
\alias{get_index}
\title{Index function}
\usage{
get_index(data = survey, grouping_var, index_var)
}
\arguments{
\item{data}{The dataset for which you will be calculating the indexes}

\item{grouping_var}{The variable for whose levels you will calculate the index}

\item{index_var}{The variable against which you will be comparing the levels of the grouping variable}
}
\value{
A dataframe with the indexes
}
\description{
This function allows you to calculate the levels of a variable in contrast to group provided by levels of another variable
}
